@@ -27,73 +27,79 @@
     "workspace-generator": "nx workspace-generator",
     "dep-graph": "nx dep-graph",
     "help": "nx help",
-    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main",
+    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2020 browser module main",
     "run:all": "node node_modules/@angular-architects/module-federation/src/server/mf-dev-server.js"
   },
   "private": true,
   "dependencies": {
-    "@angular-architects/module-federation": "^12.5.0",
-    "@angular/animations": "12.2.5",
-    "@angular/cdk": "12.2.5",
-    "@angular/common": "12.2.5",
-    "@angular/compiler": "12.2.5",
-    "@angular/core": "12.2.5",
-    "@angular/elements": "12.2.5",
-    "@angular/forms": "12.2.5",
-    "@angular/platform-browser": "12.2.5",
-    "@angular/platform-browser-dynamic": "12.2.5",
-    "@angular/router": "12.2.5",
-    "@ngrx/component": "12.4.0",
-    "@ngrx/effects": "12.4.0",
-    "@ngrx/entity": "12.4.0",
-    "@ngrx/router-store": "12.4.0",
-    "@ngrx/store": "12.4.0",
-    "@ngx-translate/core": "^13.0.0",
-    "@ngx-translate/http-loader": "^6.0.0",
-    "@nrwl/angular": "12.9.0",
-    "@nrwl/workspace": "12.9.0",
-    "@swimlane/ngx-charts": "^19.0.1",
-    "angular-oauth2-oidc": "^12.1.0",
-    "rxjs": "~6.6.7",
-    "tslib": "~2.1.0",
-    "zone.js": "~0.11.4"
+    "@angular/animations": "15.1.2",
+    "@angular/cdk": "15.1.2",
+    "@angular/common": "15.1.2",
+    "@angular/compiler": "15.1.2",
+    "@angular/core": "15.1.2",
+    "@angular/elements": "15.1.2",
+    "@angular/forms": "15.1.2",
+    "@angular/platform-browser": "15.1.2",
+    "@angular/platform-browser-dynamic": "15.1.2",
+    "@angular/router": "15.1.2",
+    "@ngrx/component": "15.1.0",
+    "@ngrx/effects": "15.1.0",
+    "@ngrx/entity": "15.1.0",
+    "@ngrx/router-store": "15.1.0",
+    "@ngrx/store": "15.1.0",
+    "@ngrx/store-devtools": "15.1.0",
+    "@ngx-translate/core": "^14.0.0",
+    "@ngx-translate/http-loader": "^7.0.0",
+    "@nrwl/angular": "15.6.3",
+    "@swimlane/ngx-charts": "^20.1.2",
+    "@webcomponents/webcomponentsjs": "^2.6.0",
+    "angular-oauth2-oidc": "^15.0.1",
+    "document-register-element": "^1.7.2",
+    "rxjs": "~7.8.0",
+    "tslib": "~2.3.1",
+    "zone.js": "~0.12.0"
   },
   "devDependencies": {
-    "@angular-architects/ddd": "^1.4.0",
-    "@angular-devkit/build-angular": "12.2.5",
-    "@angular-eslint/eslint-plugin": "12.4.1",
-    "@angular-eslint/eslint-plugin-template": "12.4.1",
-    "@angular-eslint/template-parser": "12.4.1",
-    "@angular/cli": "12.2.5",
-    "@angular/compiler-cli": "12.2.5",
-    "@angular/language-service": "12.2.5",
-    "@ngrx/schematics": "12.4.0",
-    "@ngrx/store-devtools": "12.4.0",
-    "@nrwl/cli": "12.9.0",
-    "@nrwl/cypress": "12.9.0",
-    "@nrwl/eslint-plugin-nx": "12.9.0",
-    "@nrwl/jest": "12.9.0",
-    "@nrwl/linter": "12.9.0",
-    "@nrwl/nx-cloud": "12.3.10",
-    "@nrwl/tao": "12.9.0",
-    "@types/jest": "^27.0.1",
-    "@types/node": "^16.9.1",
-    "@typescript-eslint/eslint-plugin": "^4.28.5",
-    "@typescript-eslint/parser": "^4.28.5",
-    "cpr": "^3.0.1",
-    "cypress": "^8.3.1",
+    "@angular-architects/ddd": "^3.0.0",
+    "@angular-architects/module-federation": "^15.0.3",
+    "@angular-architects/module-federation-tools": "^15.0.3",
+    "@angular-devkit/build-angular": "15.1.3",
+    "@angular-eslint/eslint-plugin": "15.1.0",
+    "@angular-eslint/eslint-plugin-template": "15.1.0",
+    "@angular-eslint/template-parser": "15.1.0",
+    "@angular/cli": "~15.1.0",
+    "@angular/compiler-cli": "15.1.2",
+    "@angular/language-service": "15.1.2",
+    "@briebug/jest-schematic": "^6.0.0",
+    "@ngrx/schematics": "15.1.0",
+    "@ngrx/store-devtools": "15.1.0",
+    "@nrwl/cli": "15.6.3",
+    "@nrwl/cypress": "15.6.3",
+    "@nrwl/eslint-plugin-nx": "15.6.3",
+    "@nrwl/jest": "15.6.3",
+    "@nrwl/linter": "15.6.3",
+    "@nrwl/workspace": "15.6.3",
+    "@types/jest": "^28.1.8",
+    "@types/node": "^16.18.11",
+    "@typescript-eslint/eslint-plugin": "^5.48.0",
+    "@typescript-eslint/parser": "^5.48.0",
+    "cypress": "^12.2.0",
     "dotenv": "^10.0.0",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-cypress": "^2.11.3",
-    "jest": "^27.1.1",
-    "jest-preset-angular": "^9.0.7",
-    "json-server": "^0.16.3",
-    "ng-packagr": "12.2.1",
-    "ngx-build-plus": "12.2.0",
-    "prettier": "^2.4.0",
-    "ts-jest": "^27.0.5",
-    "ts-node": "~10.2.1",
-    "typescript": "~4.3.5"
+    "eslint": "^8.31.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-cypress": "^2.12.1",
+    "eslint-plugin-prettier": "^3.4.1",
+    "jest": "^28.1.3",
+    "jest-environment-jsdom": "^28.1.3",
+    "jest-preset-angular": "^12.2.3",
+    "json-server": "^0.17.0",
+    "ng-packagr": "15.1.1",
+    "ngx-build-plus": "15.0.0",
+    "nx": "15.6.3",
+    "prettier": "^2.8.3",
+    "ts-jest": "^28.0.8",
+    "ts-node": "10.9.1",
+    "typescript": "~4.8.4"
   }
 }
+
