@@ -1,6 +1,6 @@
 import { DashboardPageComponent } from './dashboard-page/dashboard-page.component';
-import {Routes} from '@angular/router';
-import {HomeComponent} from './home/home.component';
+import { Routes } from '@angular/router';
+import { HomeComponent } from './home/home.component';
 
 export const APP_ROUTES: Routes = [
   {
@@ -20,4 +20,4 @@ export const APP_ROUTES: Routes = [
     path: '**',
     redirectTo: 'home'
   }
-]
+];
