@@ -1,12 +1,10 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { Directive, HostListener, Output, EventEmitter } from '@angular/core';
+import { Directive } from '@angular/core';
 
 @Directive({
   selector: 'custom-checkbox'
 })
 export class CustomCheckboxDirective {
-
   // TODO: Add @Output checkedChange to align with naming conventions
   // TODO: Add A HostListener for the custom element's changed event and trigger checkedChange
-
 }
