@@ -1,11 +1,10 @@
-import {Component} from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'navbar-cmp',
   templateUrl: 'navbar.component.html'
 })
 export class NavbarComponent {
-
   private sidebarVisible = false;
 
   sidebarToggle() {
