@@ -1,8 +1,8 @@
-import {enableProdMode} from '@angular/core';
-import {platformBrowserDynamic} from '@angular/platform-browser-dynamic';
+import { enableProdMode } from '@angular/core';
+import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
-import {AppModule} from './app/app.module';
-import {environment} from './environments/environment';
+import { AppModule } from './app/app.module';
+import { environment } from './environments/environment';
 
 // import './app/custom-checkbox/custom-checkbox';
 // import './app/hello-world/hello-world';
@@ -12,9 +12,3 @@ if (environment.production) {
 }
 
 platformBrowserDynamic().bootstrapModule(AppModule);
-
-
-
-
-
-
