@@ -1,9 +1,7 @@
-import {Component} from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'flight-booking',
   templateUrl: './flight-booking.component.html'
 })
-export class FlightBookingComponent {
-
-}
+export class FlightBookingComponent {}
