@@ -1,8 +1,8 @@
-import {Routes} from '@angular/router';
-import {FlightBookingComponent} from './flight-booking.component';
-import {FlightEditComponent} from './flight-edit/flight-edit.component';
-import {FlightSearchComponent} from './flight-search/flight-search.component';
-import {PassengerSearchComponent} from './passenger-search/passenger-search.component';
+import { Routes } from '@angular/router';
+import { FlightBookingComponent } from './flight-booking.component';
+import { FlightEditComponent } from './flight-edit/flight-edit.component';
+import { FlightSearchComponent } from './flight-search/flight-search.component';
+import { PassengerSearchComponent } from './passenger-search/passenger-search.component';
 
 export const FLIGHT_BOOKING_ROUTES: Routes = [
   {
@@ -23,5 +23,4 @@ export const FLIGHT_BOOKING_ROUTES: Routes = [
       }
     ]
   }
-
-]
+];
