@@ -3,10 +3,8 @@ import { Injectable } from '@angular/core';
 
 @Injectable()
 export class DefaultFlightCancellingService implements FlightCancellingService {
-
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    show(flightId: number): void {
-        alert('not implemented yet!');
-    }
-
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  show(flightId: number): void {
+    alert('not implemented yet!');
+  }
 }
