@@ -1,16 +1,10 @@
-import { NgModule} from '@angular/core';
+import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FlightCancellingComponent } from './flight-cancelling.component';
 
 @NgModule({
-  imports: [
-    CommonModule
-  ],
-  providers: [
-  ],
-  declarations: [FlightCancellingComponent],
-  entryComponents: []
+  imports: [CommonModule],
+  providers: [],
+  declarations: [FlightCancellingComponent]
 })
-export class FlightCancellingModule {
-
-}
+export class FlightCancellingModule {}
