@@ -2,11 +2,9 @@ import { Injectable } from '@angular/core';
 import { DefaultFlightCancellingService } from './default-flight-cancelling.service';
 
 @Injectable({
-    providedIn: 'root',
-    useClass: DefaultFlightCancellingService
+  providedIn: 'root',
+  useClass: DefaultFlightCancellingService
 })
 export class FlightCancellingService {
-
-    // show(flightId: number): void { }
-
+  // show(flightId: number): void { }
 }
