@@ -1,23 +1,21 @@
-import {Component, OnInit} from '@angular/core';
-import {ActivatedRoute} from '@angular/router';
+import { Component, OnInit } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
 
 @Component({
   selector: 'app-flight-edit',
   templateUrl: './flight-edit.component.html'
 })
 export class FlightEditComponent implements OnInit {
-  id: string;
-  showDetails: string;
+  id = '';
+  showDetails = false;
   showWarning = false;
 
-  constructor(private route: ActivatedRoute) {
-  }
+  constructor(private route: ActivatedRoute) {}
 
-  ngOnInit() {
-    this.route.params.subscribe(p => {
+  ngOnInit(): void {
+    this.route.params.subscribe((p) => {
       this.id = p['id'];
       this.showDetails = p['showDetails'];
     });
   }
-
 }
