@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-empty-function */
 /* eslint-disable @angular-eslint/no-empty-lifecycle-method */
-import {Component, OnInit} from '@angular/core';
-import {FlightService} from '@flight-workspace/flight-lib';
+import { Component, OnInit } from '@angular/core';
+import { Flight, FlightService } from '@flight-workspace/flight-lib';
 
 @Component({
   selector: 'flight-search',
@@ -9,37 +9,30 @@ import {FlightService} from '@flight-workspace/flight-lib';
   styleUrls: ['./flight-search.component.css']
 })
 export class FlightSearchComponent implements OnInit {
-
   from = 'Hamburg'; // in Germany
   to = 'Graz'; // in Austria
   urgent = false;
-
-  get flights() {
-    return this.flightService.flights;
-  }
-
   // "shopping basket" with selected flights
   basket: { [id: number]: boolean } = {
     3: true,
     5: true
   };
 
-  constructor(
-    private flightService: FlightService) {
+  constructor(private flightService: FlightService) {}
+
+  get flights(): Flight[] {
+    return this.flightService.flights;
   }
 
-  ngOnInit() {
-  }
+  ngOnInit(): void {}
 
   search(): void {
     if (!this.from || !this.to) return;
 
-    this.flightService
-      .load(this.from, this.to, this.urgent);
+    this.flightService.load(this.from, this.to, this.urgent);
   }
 
   delay(): void {
     this.flightService.delay();
   }
-
 }
