@@ -1,4 +1,4 @@
-import {Component} from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'navbar-cmp',
@@ -6,10 +6,9 @@ import {Component} from '@angular/core';
   styleUrls: ['./navbar.component.css']
 })
 export class NavbarComponent {
-
   private sidebarVisible = false;
 
-  sidebarToggle() {
+  sidebarToggle(): void {
     const body = document.getElementsByTagName('body')[0];
 
     if (this.sidebarVisible === false) {
