@@ -1,32 +1,24 @@
-import {CommonModule} from '@angular/common';
-import {NgModule} from '@angular/core';
-import {ModuleWithProviders} from '@angular/core';
-import {CityPipe} from './pipes/city.pipe';
+import { CommonModule } from '@angular/common';
+import { ModuleWithProviders, NgModule } from '@angular/core';
+import { CityPipe } from './pipes/city.pipe';
 
 @NgModule({
-  imports: [
-    CommonModule
-  ],
-  declarations: [
-    CityPipe,
-  ],
-  exports: [
-    CityPipe
-  ]
+  imports: [CommonModule],
+  declarations: [CityPipe],
+  exports: [CityPipe]
 })
 export class SharedModule {
   static forRoot(): ModuleWithProviders<SharedModule> {
     return {
       ngModule: SharedModule,
       providers: []
-    }
+    };
   }
 
   static forChild(): ModuleWithProviders<SharedModule> {
     return {
       ngModule: SharedModule,
       providers: []
-    }
+    };
   }
-
 }
