@@ -3,7 +3,7 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'flight-workspace-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss'],
+  styleUrls: ['./app.component.scss']
 })
 export class AppComponent {
   title = 'passenger';
