@@ -1,4 +1,4 @@
-export {FlightLibModule} from './lib/flight-lib.module';
-export {Flight} from './lib/models/flight';
-export {FlightService} from './lib/services/flight.service';
-export {FakeFlightService} from './lib/services/fake-flight.service';
+export { FlightLibModule } from './lib/flight-lib.module';
+export { Flight } from './lib/models/flight';
+export { FlightService } from './lib/services/flight.service';
+export { FakeFlightService } from './lib/services/fake-flight.service';
