@@ -1,17 +1,14 @@
-import {CommonModule} from '@angular/common';
-import {ModuleWithProviders, NgModule} from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ModuleWithProviders, NgModule } from '@angular/core';
 
 @NgModule({
-  imports: [
-    CommonModule
-  ]
+  imports: [CommonModule]
 })
 export class FlightLibModule {
   static forRoot(): ModuleWithProviders<FlightLibModule> {
     return {
       ngModule: FlightLibModule,
       providers: []
-    }
+    };
   }
 }
-
