@@ -1,8 +1,7 @@
-
 export interface Flight {
-  id: number;   
+  id: number;
   from: string;
   to: string;
-  date: string; 
+  date: string;
   delayed: boolean;
 }
