@@ -1,4 +1,3 @@
-
 export abstract class LoggerConfig {
-    enableDebug?: boolean;
+  enableDebug?: boolean;
 }
