@@ -1,8 +1,4 @@
 import { Injectable } from '@angular/core';
-import { BehaviorSubject } from 'rxjs';
-
 
 @Injectable({ providedIn: 'root' })
-export class EditFacade {
-
-}
+export class EditFacade {}
