@@ -2,6 +2,6 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 @NgModule({
-  imports: [CommonModule],
+  imports: [CommonModule]
 })
 export class PassengerDomainModule {}
