@@ -2,9 +2,9 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
-import { ReactiveComponentModule } from '@ngrx/component';
 import { SearchComponent } from './search.component';
 import { PassengerDomainModule } from '@flight-workspace/passenger/domain';
+import { LetModule } from '@ngrx/component';
 
 @NgModule({
   imports: [
@@ -17,10 +17,10 @@ import { PassengerDomainModule } from '@flight-workspace/passenger/domain';
         component: SearchComponent
       }
     ]),
-    ReactiveComponentModule,
-    PassengerDomainModule
+    PassengerDomainModule,
+    LetModule
   ],
   declarations: [SearchComponent],
-  exports: [SearchComponent],
+  exports: [SearchComponent]
 })
 export class PassengerFeatureSearchModule {}
